wikimedia chess pieces: Use new recolouring system, change colours
We replace red/yellow/green with red/yellow/light-blue/purple.

This provides, additionally, the missing colour variants.

Clare helpfully uploaded a white ship piece to Wikimedia, so we let
media-scraper download that.

The coloured pieces in Penultima are replaced by the ones we provide
here.  Mostly this is just a name change, but "g" are supplanted by
"blue".

The chess-w-* files are renamed.  The others are now obsolete.

Signed-off-by: Clare Boothby
Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>