hidden prep: Reorganise PieceUpdateFromOp
We are going to want to allow an op to give different answers to
different clients.  The representation from an op is going to be
either basically what we have now, or a version where the
PieceUpdateOp is different for each player.

Prepare for this with some struct and type shuffling.  We laave a
convenience Into impl to allow ops to easily handle the simple case.

No functional change.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>