config: Allow OTTER_LOG to (completely) override config
If the env var is supplied and results in a nonempty log spec, we
replace the log config completely.

This is quite a lot of faff because no-one likes to give us the right
interfaces.  In particular flexi_logger wants to give and receive
(byte) strings.

This is not the final state; we want to merge them, really.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>