Makefile: Drop use of wasm-pack
I'm told on Discord that this is obsolete.  (The docs all still
mention it, though.)

Anyway, the only thing it was actually doing for us was obtaining and
running wasm-bindgen.  We can do that ourselves.

At one time we were using it to get and run wasm-opt, saving 5k out of
50k in the wasm file, but that was broken so we disabled it a while
ago.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>