Prevent use of organise function when rotated
Add the proof token to the organise function, and check it in hand.

This only currently makes any difference for player labels, because
the other kinds of hand are unrotateable from load time.

Because we don't update uos on rotation, the player can try to
organise a rotated player label, but they will get an error.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>