script: When lowering, avoid raising
It can happen that the z of the piece which sets the z_top for the end
of the plan (with z_bot == null) will cause the first lowered piece to
actually be raised.

This is particularly troublesome if the piece is an occulting one,
which the server will not allowe to be raised.

Handle this case specially: if necessary, set the z_top to top piece
of that plan entry.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>