Extend InertPieceTrait to support multiple faces
We want this because we want to use InertPiece for dice and a die can
have multiple faces, which we want to specify as a single piece.  Ie,
the die face images will be an inert version of the die.

No functional change right now - we just move the default() call out
from the trait impls to each of the call sites.  In the shapelib
impl (which is the only impl right now) we pass through the face to
the rendering code (which lives in the main PieceTrait impl).

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>