Implement InertPieceTrait for SimpleShape
Right now, this allows simple circles and rectangles to be used as
backs (for occultation).  It will also allow them to be used for dice,
when we have those.

It is not convenient to provide a test for this because occultation
methods must be provided in libraries, and we don't have a test
library.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>