dice: Tidy remprop handling
This macro was overkill.  It existed mostly to avoid uselessly
comparing the remaining duration with 0 and storing None when it was
there already.  That's a daft microoptimisation.

Doing it this way instead lets us only do the remprop calculation when
it's wanted, and it will let us provide a way to set cooldown_expires
to None if and when we have &mut State and can see it's timed out.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>