dice: Break out OccultSpec (breaking change)
Earlier I wrote: We want to add an ilk here.

But this is no longer going to be true, because we are getting rid of
ilk mixing for dice.  Nevertheless, this change is going in a roughly
good direction in spec API terms since it will allow us to add other
things to occult.

Notably, if we allow composition of a die image from a base and faces,
or if we allow specifying an occulted face image separately from the
face images.

And `occult_label` is just a bad pattern.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>