dice: Reorganise occultation handling
I was trying to support ilks properly.  This code is in fact going to
be considerably simplified.

But it will still make sense to split the interpretation of the spec
from the assembly of the Arc<dyn InertPieceImage> for the occulted
view.  That will make it possible for the spec interpretation to
provide an occulted view in more situations.

In fact it appears that these plans are going to be postponed and not
done right now, but I'm keeping this refactoring.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>