Elide some fields from save game state where we can
This saves writing out the field names as well as just the vslues.

We would like to do this for the updates sent to the client, but that
would really depend on the JS nullable coalesce operator "??" which is
not available the version of typescript I'm using, nor our ES
compatibility target.  Newer typescript would do a polyfill but that
maybe wouldn't be super-fast?  So postpone that.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>