currency: Abolish min_unit
We can instead just rely on the players to Not Do That, or to tell
each other off.  It's easy to recover from, too: just move money until
the amounts are right.

This also avoids any question of what to do if people try to combine
banknotes with the same currency but different min_qty.  (GCD?)

Suggested-by: Gareth Taylor
Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>