svg size handling: Introduce new size parser
We want to get the size out of the svg on load, to make the library
spec size attribute optional (and to make things less confusing).
So start by add code to parse size out of an SVG.

But in fact it turns out we have one of these in bundle processing
already.  The need to be combined.  But for now, commit what we have.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>