mformat 1: Have FaceTransform::from_group_mf1 take the GroupData
This gives access to mformat, which is useful for error
reporting (since it allows us to use the standard "can't do that in
this mf version" constructor).

Of course this is an _mf1 function so it doesn't need the actual
value - it could perhaps have made do with a proof token of const
generic type, or something.  But I tried that and it was a palaver.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>