mformat 1: Drop the size parameter from load_mf1
This was introduced in
  size handling: Plumb svg size through to OutlineDefn
and it seemed a good idea at the time.  But when I wrote that I hadn't
decided that these entrypoints would be renamed and kept, rather than
start to contain conditionals.

The mf1 versions can't depend on the size, since in format 1 the SVG's
actual size wasn't used.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>