MaterialsFormatVersionError: Improve messages
Well, I managed to produce these.  But the output was a bit ugly.

I think the stuff in Other about a "reasonable integer" is covered by
the specific strings we pass in the integer conversions, and there's
going to be a possibility of completely different kinds of problem:

Use of a format default in a builtin shape library is going to become
an error.  And we don't want to add a variant for it - that's
overkill.  So we'll reuse Other, and this message is not right.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>